backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==1.2.2
firecrawl-py==1.5.0
gunicorn==21.2.0
